Stabilize control IPC and desktop ownership
This patch hardens the control IPC host with Unix flock-based ownership checks, stale socket cleanup, and defensive handling for malformed host responses and invalid event payloads. It also centralizes blocking SQLite/filesystem work behind a shared module helper and migrates legacy host pushes to domain-event topics for live, plugin, and points updates. The dev launcher now refuses stale desktop owners so a dead or stale app cannot be mixed with a new Vite session, and the points UI refreshes the leaderboard when awards arrive for new viewers instead of silently dropping them.

+++ crates/tiktools-tiktok/src/lib.rs
@@ -128,7 +128,7 @@ pub enum TikTokError {
 pub enum ClientEvent {
     Connected(ConnectionInfo),
     Event(events::TikToolsEvent),
-    Reconnecting { attempt: u32, delay_ms: u64 },
+    Reconnecting { attempt: u64, delay_ms: u64 },
     Disconnected { reason: String },
     Error { phase: ErrorPhase, message: String },
 }
